backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
haystack-ai>=2.7.0
chromadb>=1.0.0
openai>=1.50.0
python-dotenv==1.0.1
werkzeug==3.1.3
requests==2.32.3
chroma-haystack>=0.20.0
